Change structure of conversion folder.
